% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_ASAP.R
\name{PlotSimASAP}
\alias{PlotSimASAP}
\title{PlotSimASAP}
\usage{
PlotSimASAP(wd, asap.name, whichsim, od = file.path(wd, "sim"),
  save.plots = FALSE)
}
\arguments{
\item{wd}{directory where original ASAP run is located}

\item{asap.name}{base name of original dat file (without the .dat extension)}

\item{whichsim}{vector of numbers for which simulated data sets to plot}

\item{od}{location where new data sets are saved (defaults to sim subdirectory of wd)}

\item{save.plots}{will save png files if TRUE (default = FALSE)}
}
\description{
Comparison plots of SSB, Freport, and Recruitment from SimASAP runs with true values.
}
