% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ASAP.R
\name{SimASAP}
\alias{SimASAP}
\title{SimASAP}
\usage{
SimASAP(wd, asap.name, nsim, od = file.path(wd, "sim"),
  runflag = FALSE)
}
\arguments{
\item{wd}{directory where original ASAP run is located}

\item{asap.name}{base name of original dat file (without the .dat extension)}

\item{nsim}{number of simulated data sets to create}

\item{od}{location where new data sets are saved (defaults to sim subdirectory of wd)}

\item{runflag}{flag to run ASAP on new data sets (TRUE) or not (FALSE) (default = FALSE)}
}
\description{
Generate a number of ASAP data sets with observation error from one already run and optionally run them.
}
